[solver]: fix 2d shallow water benchmarks

+++ Application/SWE/SWE2d/@SWEAbstract2d/matImposeBoundaryCondition.m
@@ -1,7 +1,7 @@
 function [ fM, fP ] = matImposeBoundaryCondition( obj, edge, nx, ny, fM, fP, fext )
 
 % impose boundary condition
-[ fP ] = mxImposeBoundaryCondition( obj.gra, nx, ny, fP, fext, edge.ftype );
+[ fP ] = mxImposeBoundaryCondition( obj.gra, nx, ny, fP, fext, int8( edge.ftype ) );
 
 % hydrostatic reconstruction
 [ fM, fP ] = mxHydrostaticReconstruction( obj.hmin, fM, fP );
